Rewrite cvcreate with Jinja2 and pydantic
Current implementation of cvcreate is a bit of a mess.
So I rewrote it, but tried to use some known tools along the way to make it better.

Jinja
-----

Jinja is a known templating tool spun out of Django. With some (recommended) modified defaults it allows us to create latex templates with valid latex syntax.
In practice we use `\VAR{ ... }` to inserts a variable, and `\BLOCK{ ... }` is an evaluation block.
In addition to standardizing the format, it has allowed us to move processing blocks like string-interpolation and conditional blocks into the template format.

Pydantic
--------

Verifying what the user brings to the table is not done much right now.
Pydantic improves on this by validating input against annotation.

YAML source scheme
------------------

As noted in #42, the format we are currently using is problematic. This
Pydantic+Jinja solution requires us to change it a bit.
See new `example.yaml` to see the new format.

New flag features
-----------------

The new parser (which is mostly a copy of the old one, with less bloat),
has removed some old flags (like logo position), and instead introduces:

`--logo-image`: use a custom logo image (or personal image)
`--footer-image`: use a custom footer image (set it to `footer_de` for Vinz)
`--font-size`: use non-default font-size.

These are added as a proof-of-concept. We can easily add lots of more stuff later.